fix: the pulse stepped through four brightnesses instead of swelling
Fx.Glow declared its alpha as an int, but that value only ever reaches GDI+
through Matrix33, which is float the whole way. The integer was therefore
pure loss, and it landed exactly where it hurt: PillBar's two breathing
glows were computed as (int)(16 * strength * lit) and (int)(13 * strength *
lit), which across a whole breath is FOUR distinct values and THREE. A slow
swell was rendering as a handful of visible steps.

Measured at frame rate: the wide glow's argument went from 4 levels per
breath to continuous, and a pixel mid-fill now takes 24-26 distinct values
across a breath instead of stepping with the glow. What is left is the
8-bit alpha of the fill itself, which is GDI+'s floor, not ours.

All seventeen Glow callers pass int literals and are unaffected.

Downloads breathe now too, on the same rule as media: a download 40% done
and one stalled at 40% are the same still picture, and the pill's bar is
what you glance at to tell them apart. A paused one holds still.

+++ src/Halo.App/Widgets/Fx.cs
@@ -87,8 +87,13 @@ internal static class Fx
     // very soft radial wash of the accent, clipped to the pill shape
     // (flat top flush to the screen edge + rounded bottom — same as LayeredNotch.PillPath;
     // an all-corners clip leaves a dark crescent at the top corners)
+    // alpha is a float, and deliberately so: it only ever reaches GDI+ through Matrix33 below, which is
+    // float the whole way. Declaring it int quantised every caller AT THE CALL SITE - PillBar's breathing
+    // glows landed on (int)(16*0.5*lit) and (int)(13*0.5*lit), which is four levels and three, so a pulse
+    // meant to be a slow swell stepped visibly through a handful of brightnesses. Nothing about the drawing
+    // needed the integer; the cast was the banding.
     public static void Glow(Graphics g, int w, int h, float fade, float cx, float cy,
-        float rx, float ry, int alpha, Color accent)
+        float rx, float ry, float alpha, Color accent)
     {
         if (accent == White || fade <= 0.01f) return;
         using var clip = PillClip(w, h);
@@ -185,7 +190,7 @@ internal static class Fx
             var oldG = g.Clip;
             g.SetClip(new RectangleF(0, 0, fill, h), CombineMode.Intersect);
             Glow(g, w, h, fade, fill * 0.45f, h * 0.44f, Math.Max(fill, h * 1.2f), h * 1.9f,
-                 (int)(16 * strength * lit), accent);
+                 16 * strength * lit, accent);
             g.Clip = oldG;
         }
 
@@ -258,7 +263,7 @@ internal static class Fx
             var oldG = g.Clip;
             g.SetClip(new RectangleF(0, 0, fill, h), CombineMode.Intersect);
             Glow(g, w, h, fade, fill, h / 2f, h * 1.1f, h * 1.45f,
-                 (int)(13 * strength * lit), accent);
+                 13 * strength * lit, accent);
             g.Clip = oldG;
         }
     }
